[PWGCF] FemtoUniverse: Fix bug in badpoint fraction calculation (#15550)

+++ PWGCF/FemtoUniverse/Core/FemtoUniverseDetaDphiStar.h
@@ -877,7 +877,7 @@ class FemtoUniverseDetaDphiStar
         badpoints++;
       }
     }
-    return badpoints / entries;
+    return (static_cast<float>(badpoints) / static_cast<float>(entries));
   }
 
   // Get particle charge from mask
